requirements.txt
streamlit
pillow
face_recognition
numpy
scipy
matplotlib
plotly==5.3.1
opencv==4.5.3